Cash transferred amount changed to decimal and is configurable new\old

+++ App/EzService/EzService/EzServiceImplementation.cs
@@ -255,7 +255,7 @@ namespace EzService {
 			return Execute(typeof (ApprovedUser), customerId, loanAmount);
 		} // ApprovedUser
 
-		public ActionMetaData CashTransferred(int customerId, int amount) {
+		public ActionMetaData CashTransferred(int customerId, decimal amount) {
 			return Execute(typeof (CashTransferred), customerId, amount);
 		} // CashTransferred
