+++ Backend/Strategies/MailStrategies/CashTransferred.cs
@@ -7,7 +7,7 @@ namespace EzBob.Backend.Strategies.MailStrategies {
 	public class CashTransferred : AMailStrategyBase {
 		#region consturctor
 
-		public CashTransferred(int customerId, int amount, AConnection oDB, ASafeLog oLog) : base(customerId, true, oDB, oLog) {
+		public CashTransferred(int customerId, decimal amount, AConnection oDB, ASafeLog oLog) : base(customerId, true, oDB, oLog) {
 			this.amount = amount;
 		} // constructor
 
@@ -37,6 +37,6 @@ namespace EzBob.Backend.Strategies.MailStrategies {
 
 		#endregion method SetTemplateAndSubjectAndVariables
 
-		private readonly int amount;
+		private readonly decimal amount;
 	} // class CashTransferred
 } // namespace
